Add MapAt failure tests...
and split out tests better.
